some changes

+++ 0x08-recursion/main.h
@@ -5,4 +5,4 @@ int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
 int _sqrt_recursion(int n);
-int getroot(int n, int x);
+int getroot(int n, long x);
